move tests. remove attribute related tests
